requirements.txt
asgiref==3.3.4
Django==3.2
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
gunicorn
django-heroku
django-cloudinary-storage
python-decouple